+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 This is the most recent Changelog. It also contains some of the changes inside the >>>UPCOMING<<< Versions (most of the time). This is for 1.7.10 btw.
 
+6.1.0-plus: (Not released yet)
+[ADDED] zh_CN language file is in default config pack.Thanks to translation groupware.
+[REMOVED] Print_Missing_I18n_Key config is removed due to unstable.
+
 6.0.5-vd2:
 [FIXED] The crucible heat loss is more smooth.
 [ADDED] Waila Compat, in WIP.